break up auto-materialize toys between lazy and eager
