+++ python_modules/dagster-test/dagster_test/toys/auto_materializing/repo_2.py
@@ -6,7 +6,6 @@ from dagster import (
     asset,
     repository,
 )
-from dagster._core.definitions.freshness_policy import FreshnessPolicy
 
 eager_downstream_1_source = SourceAsset(AssetKey(["eager_downstream_1"]))
 
@@ -26,27 +25,6 @@ def eager_downstream_3(eager_downstream_1):
 @asset(auto_materialize_policy=AutoMaterializePolicy.eager())
 def eager_downstream_4(eager_downstream_2, eager_downstream_3):
     return eager_downstream_2 + eager_downstream_3
-
-
-lazy_downstream_1_source = SourceAsset(AssetKey(["lazy_downstream_1"]))
-
-
-@asset(auto_materialize_policy=AutoMaterializePolicy.lazy())
-def lazy_downstream_2(lazy_downstream_1):
-    return lazy_downstream_1 + 2
-
-
-@asset(auto_materialize_policy=AutoMaterializePolicy.lazy())
-def lazy_downstream_3(lazy_downstream_1):
-    return lazy_downstream_1 + 3
-
-
-@asset(
-    auto_materialize_policy=AutoMaterializePolicy.lazy(),
-    freshness_policy=FreshnessPolicy(maximum_lag_minutes=1440),
-)
-def lazy_downstream_4(lazy_downstream_2, lazy_downstream_3):
-    return lazy_downstream_2 + lazy_downstream_3
 
 
 ### Partitioned ###
@@ -73,28 +51,6 @@ def eager_downstream_4_partitioned(eager_downstream_2_partitioned, eager_downstr
     return eager_downstream_2_partitioned + eager_downstream_3_partitioned
 
 
-lazy_downstream_1_source_partitioned = SourceAsset(AssetKey(["lazy_downstream_1_partitioned"]))
-
-
-@asset(partitions_def=daily_partitions_def, auto_materialize_policy=AutoMaterializePolicy.lazy())
-def lazy_downstream_2_partitioned(lazy_downstream_1_partitioned):
-    return lazy_downstream_1_partitioned + 2
-
-
-@asset(partitions_def=daily_partitions_def, auto_materialize_policy=AutoMaterializePolicy.lazy())
-def lazy_downstream_3_partitioned(lazy_downstream_1_partitioned):
-    return lazy_downstream_1_partitioned + 3
-
-
-@asset(
-    partitions_def=daily_partitions_def,
-    auto_materialize_policy=AutoMaterializePolicy.lazy(),
-    freshness_policy=FreshnessPolicy(maximum_lag_minutes=1440),
-)
-def lazy_downstream_4_partitioned(lazy_downstream_2_partitioned, lazy_downstream_3_partitioned):
-    return lazy_downstream_2_partitioned + lazy_downstream_3_partitioned
-
-
 @repository
 def auto_materialize_repo_2():
     return [
@@ -102,16 +58,8 @@ def auto_materialize_repo_2():
         eager_downstream_3,
         eager_downstream_1_source,
         eager_downstream_4,
-        lazy_downstream_2,
-        lazy_downstream_3,
-        lazy_downstream_1_source,
-        lazy_downstream_4,
         eager_downstream_2_partitioned,
         eager_downstream_3_partitioned,
         eager_downstream_1_source_partitioned,
         eager_downstream_4_partitioned,
-        lazy_downstream_2_partitioned,
-        lazy_downstream_3_partitioned,
-        lazy_downstream_1_source_partitioned,
-        lazy_downstream_4_partitioned,
     ]
